Move errors to types folder
