+++ lib/types/errors.ts
