+++ pages/apply/verify.tsx
@@ -10,7 +10,7 @@ import { useAppDispatch, useAppSelector } from '../../lib/store/hooks';
 import { FormData } from '../../lib/types/form';
 import { FormID, getFormData } from '../../lib/utils/form-data';
 import ErrorSummary from '../../components/errors/error-summary';
-import { Errors } from '../../lib/utils/errors';
+import { Errors } from '../../lib/types/errors';
 import { scrollToError } from '../../lib/utils/scroll';
 import { confirmVerifyCode, createVerifyCode } from '../../lib/store/auth';
